A new example that trains a small model on small MNIST images.
This example also showcases how to use the alternating projections API
and is a full implementation of the VIKING algorithm.

Added a test that runs this example, but only if the environment
variable "RUNALL" is set, since this example has external
dependencies (both Python and downloaded data) and takes a long time
to run.